image/requirements.txt
bcrypt==4.2.0
blinker==1.8.2
cachetools==5.5.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
click==8.1.7
cryptography==43.0.1
Flask==3.0.3
google-api-core==2.19.2
google-auth==2.34.0
google-cloud-pubsub==2.23.0
googleapis-common-protos==1.65.0
grpc-google-iam-v1==0.13.1
grpcio==1.66.1
grpcio-status==1.66.1
idna==3.8
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
proto-plus==1.24.0
protobuf==5.28.0
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
PyNaCl==1.5.0
requests==2.32.3
rsa==4.9
urllib3==2.2.2
Werkzeug==3.0.4
